feat: add conditional for docker stages

+++ Jenkinsfile
@@ -36,19 +36,23 @@ pipeline {
         stage("Docker compose") {
             steps {
                 script {
-                    sh(script: "docker compose down")
-                    sh(script: "docker compose up -d --build")
+                    if (env.BRANCH_NAME == "main" || env.BRANCH_NAME == "master") {
+                        sh(script: "docker compose down")
+                        sh(script: "docker compose up -d --build")
+                    }
                 }
             }
         }
 
         stage("Upload to Docker Hub") {
             steps {
-                sh(script: "docker login -u inkeister -p Ink@0346333767")
-                sh(script: "docker tag dbf_api inkeister/dbf_api:${project_version}")
-                sh(script: "docker push inkeister/dbf_api:${project_version}")
-                sh(script: "docker tag dbf_api inkeister/dbf_api:latest")
-                sh(script: "docker push inkeister/dbf_api:latest")
+                if (env.BRANCH_NAME == "main" || env.BRANCH_NAME == "master") {
+                    sh(script: "docker login -u inkeister -p Ink@0346333767")
+                    sh(script: "docker tag dbf_api inkeister/dbf_api:${project_version}")
+                    sh(script: "docker push inkeister/dbf_api:${project_version}")
+                    sh(script: "docker tag dbf_api inkeister/dbf_api:latest")
+                    sh(script: "docker push inkeister/dbf_api:latest")
+                }
             }
         }
     }
